+++ src/config/languages/spanishContent.js
@@ -17,27 +17,25 @@ const spanishContent = {
     },
     homePage: {
         banner: {
-            title: 'Hola world, I am Gabi Polanco-Ferreyra',
+            title: 'Hola world, Soy Gabi Polanco-Ferreyra',
             subtitle:
-                'My background in law allowed me to develop my empathic view of the world, and my limitless curiosity led me to study different design programs in a self-taught and formal way.',
-            subtitle2: 'The result: Gabi,  ',
-            subtitle3: 'new UX/UI designer who wants to be part of the ',
-            subtitle4: "I'm driven by the desire to make this world more accessible, fair, and democratic.",
-            subtitleBold1: '(very empathic)',
-            subtitleBold2: 'digital revolution'
+                'Mi background en Derecho me permitió desarrollar mi visión empática del mundo, y mi curiosidad sin límites, me llevo a estudiar diferentes programas de diseño de forma autodidacta y formalmente.',
+            subtitle2: 'El resultado: Gabi',
+            subtitle3: 'una nueva diseñadora UX/UI que quiere ser parte',
+            subtitle4: 'Me mueve el deseo de hacer este mundo uno más accesible, justo y democrático.',
+            subtitleBold1: '(muy empática)',
+            subtitleBold2: 'de la revolución digital'
         },
         selectedWorks: {
-            title: 'Selected works',
-            firstProjectTitle: 'Rito',
-            firstProjectSubtitle: 'web/app editorial project',
-            secondProjectTitle: 'Shaka App',
-            secondProjectSubtitle: 'mobile wellness app'
+            title: 'Selección de trabajos',
+            caseStudiesTitle: 'Case Studies',
+            webDesignTitle: 'Diseño web'
         },
         about: {
             title: 'about me',
             content:
-                'Soy Gabi Polanco-Ferreyra and I come from a colorful background, graduated from Law School, did a master’s in Mediation and Conflict Resolutions, worked in government for many years before I dedicated myself to being a full-time traveler around the world for a few months. I tried a lot of other jobs before committing to my creative talents full-time.I love to travel, see the world, camp, hike mountains, and meet people (really). My dream is to sail around the world one day.',
-            contentHighlight: 'I am on my creative-human journey, and ME ENCANTA.',
+                'Soy Gabi Polanco-Ferreyra y vengo de un pasado profesional variado, me gradué en la Escuela de Derecho, hice un máster en Mediación y Resolución de Conflictos, trabajé en el gobierno durante muchos años antes de dedicarme a ser una viajera a tiempo completo alrededor del mundo durante unos meses. Tuve muchos otros trabajos antes de decidir dedicarme a mis talentos creativos como profesión. Me encanta viajar, ver el mundo, acampar, escalar montañas y conocer gente (de verdad). Mi sueño es dar la vuelta, en velero, al mundo algún día.',
+            contentHighlight: 'Estoy transitando mi viaje creativo and ME ENCANTA.',
             linkedinLinkText: 'Linkedin',
             linkedinLink: 'https://www.linkedin.com/in/gabriela-polanco-ferreyra-b5b25717a/',
             mediumLinkText: 'Medium',
@@ -46,7 +44,7 @@ const spanishContent = {
             cvLink: '/resume.pdf',
             behanceLinkText: 'Behance',
             behanceLink: 'https://www.behance.net/gabrielapolanco2',
-            askForAvailability: 'ask for availability'
+            askForAvailability: 'Hablemos'
         }
     },
     landingPage: {
@@ -186,7 +184,8 @@ const spanishContent = {
         researchLabel: 'Research',
         visualConceptLabel: 'Visual Concept',
         prototypeLabel: 'Prototype',
-        checkOnBehanceLink: 'CHECK OUT THE FULL CASE STUDY IN BEHANCE.'
+        checkHereLink: 'CHECK OUT THE FULL CASE STUDY',
+        hereTitle: 'HERE'
     },
     settingsPage: {
         menu: {