+++ src/data/rito.js
@@ -1,3 +1,10 @@
+import rito1 from 'assets/images/rito1.jpg';
+import rito2 from 'assets/images/rito2.jpg';
+import rito3 from 'assets/images/rito3.jpg';
+import rito4 from 'assets/images/rito4.jpg';
+import rito5 from 'assets/images/rito5.jpg';
+import rito6 from 'assets/images/rito1.jpg';
+
 const rito = {
     id: 'rito',
     behanceLink: 'https://behance.net',
@@ -10,30 +17,223 @@ const rito = {
         ],
         scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
         categories: 'Website/Mobile ',
-        brief: 'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
-        research: [
-            'Starting our creative process by choosing an already created user persona (Paula) allowed us to think about this editorial project, which mutated and evolved from the beginning excitingly and organically. By having specific characteristics about Paula, we could orient our research.',
-            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
-            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.'
-        ],
-        visualConcept:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        prototype:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.'
+        post: [
+            {
+                sectionNumber: '01',
+                id: 'brief',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito1, alt: 'Rito 1' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '02',
+                id: 'research',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'images-slide',
+                        content: [
+                            { src: rito2, alt: 'Rito 2' },
+                            { src: rito3, alt: 'Rito 3' }
+                        ]
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'We interviewed five people and this is what we learned from them.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito4, alt: 'Rito 4', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'On the other hand, the surveys allow us to learn that magazines are a product that is not consumed daily, but more sporadically during the week. Also, 100% of the time, they are accessed by recommendation on social networks. Finally, people consume them because they look for inspiration and recommendations.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'In this stage of seeking to empathize with our users, we opened the door to Latin American art. We decided to create RITO to contribute to the visibility and democratization of emerging artists from Argentina and Brazil, countries that share territorial borders, proximity, and a long history of cooperation despite not sharing the same language.',
+                        size: '400',
+                        weight: '600',
+                        style: 'normal'
+                    }
+                ]
+            },
+            {
+                sectionNumber: '03',
+                id: 'visualConcept',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We choose a  colorful palette since colors are representative of Latin America.Our name was born after brainstorming a lot. We thought about our interviews, what we heard about them, and reflected on what we understood about art.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito5, alt: 'Rito 5', size: 'auto' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '04',
+                id: 'prototype',
+                content: [
+                    {
+                        type: 'video',
+                        content: { src: rito6, alt: 'Rito 6' }
+                    }
+                ]
+            }
+        ]
     },
     es: {
         subtitle: 'web/app editorial project',
-        overview:
-            'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers.',
-        focus: 'Experience and Interface Design',
+        overview: [
+            'Project created from scratch, in 4 days, in collaboration with Tadeau Carabias.',
+            'The challenge was to create a responsive online platform for a magazine to meet the needs and goals of users. The goal was to deliver a consistent, branded, and engaging experience and our main focus were on UI Design.'
+        ],
+        scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
         categories: 'Website/Mobile ',
-        brief: 'Rito is a collaborative project,  done with Tadeu Carabias in  is a  magazine oriented to a young audience. Our public intended to be an audience who enjoys knowing new and emergent artists, who like to know what happens in the artistic environment, is aware of new trends, and receives recommendations on new artistic creations to continue feeding their artistic interest spirit.',
-        research:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        visualConcept:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
-        prototype:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.'
+        post: [
+            {
+                sectionNumber: '01',
+                id: 'brief',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito1, alt: 'Rito 1' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '02',
+                id: 'research',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'images-slide',
+                        content: [
+                            { src: rito2, alt: 'Rito 2' },
+                            { src: rito3, alt: 'Rito 3' }
+                        ]
+                    },
+                    {
+                        type: 'paragraph',
+                        content: 'We interviewed five people and this is what we learned from them.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito4, alt: 'Rito 4', size: 'auto' }
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'On the other hand, the surveys allow us to learn that magazines are a product that is not consumed daily, but more sporadically during the week. Also, 100% of the time, they are accessed by recommendation on social networks. Finally, people consume them because they look for inspiration and recommendations.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'paragraph',
+                        content:
+                            'In this stage of seeking to empathize with our users, we opened the door to Latin American art. We decided to create RITO to contribute to the visibility and democratization of emerging artists from Argentina and Brazil, countries that share territorial borders, proximity, and a long history of cooperation despite not sharing the same language.',
+                        size: '400',
+                        weight: '600',
+                        style: 'normal'
+                    }
+                ]
+            },
+            {
+                sectionNumber: '03',
+                id: 'visualConcept',
+                content: [
+                    {
+                        type: 'paragraph',
+                        content:
+                            'We choose a  colorful palette since colors are representative of Latin America.Our name was born after brainstorming a lot. We thought about our interviews, what we heard about them, and reflected on what we understood about art.',
+                        size: '200',
+                        weight: '400',
+                        style: 'normal'
+                    },
+                    {
+                        type: 'image',
+                        content: { src: rito5, alt: 'Rito 5', size: 'auto' }
+                    }
+                ]
+            },
+            {
+                sectionNumber: '04',
+                id: 'prototype',
+                content: [
+                    {
+                        type: 'video',
+                        content: { src: rito6, alt: 'Rito 6' }
+                    }
+                ]
+            }
+        ]
     }
 };