+++ src/pages/projects/Sections.js
@@ -4,20 +4,18 @@ import clsx from 'clsx';
 
 import { Grid, makeStyles, Typography } from '@material-ui/core';
 
-import rito1 from 'assets/images/rito1.jpg';
-import rito2 from 'assets/images/rito1.jpg';
-import rito3 from 'assets/images/rito1.jpg';
-import rito4 from 'assets/images/rito1.jpg';
-import rito5 from 'assets/images/rito1.jpg';
-import rito6 from 'assets/images/rito1.jpg';
-
 const useStyles = makeStyles((theme) => ({
-    firstSectionImage: {
-        height: '60vh'
+    checkOutLink: {
+        fontSize: '2rem',
+        color: theme.palette.primary.main,
+        fontWeight: 700,
+        padding: '32px 64px',
+        [theme.breakpoints.up('md')]: {
+            fontSize: '3rem'
+        }
     },
-    fourthSectionImage: {
-        height: '80vh',
-        backgroundColor: '#C4C4C4'
+    hereLink: {
+        textDecoration: 'underline!important'
     },
     image: {
         display: 'flex',
@@ -27,16 +25,26 @@ const useStyles = makeStyles((theme) => ({
             height: '100%'
         }
     },
-    secondSectionImage: {
-        height: '60vh',
-        backgroundColor: '#C4C4C4'
+    imageSizeAuto: {
+        height: 'auto',
+        width: '100%',
+        '&>img': {
+            width: '100%'
+        }
+    },
+    imageSizeNormal: {
+        height: '60vh'
+    },
+    imagesSlideWrapper: {
+        display: 'flex',
+        gap: '2rem'
     },
     section: {
         position: 'relative',
         paddingTop: '80px!important'
     },
     sectionContent: {
-        marginBottom: '2rem'
+        margin: '1rem 0'
     },
     sectionIndex: {
         [theme.breakpoints.up('lg')]: {
@@ -45,216 +53,108 @@ const useStyles = makeStyles((theme) => ({
             left: 0,
             transform: 'translateX(-100%)'
         }
-    },
-    thirdSectionImage: {
-        height: '40vh',
-        backgroundColor: '#C4C4C4'
     }
 }));
-
-const images = {
-    rito: {
-        image1: rito1,
-        image2: rito2,
-        image3: rito3,
-        image4: rito4,
-        image5: rito5,
-        video1: rito6
-    },
-    'shaka-app': {
-        image1: '',
-        image2: '',
-        image3: '',
-        image4: '',
-        image5: '',
-        image6: '',
-        image7: '',
-        image8: '',
-        image9: ''
-    }
-};
 
 const Sections = ({ classes, ...props }) => {
     const internalClasses = useStyles();
 
-    const project = useSelector(({ entities }) => entities.projects.currentProject);
+    const project = useSelector(({ entities }) => entities.projects.currentProject?.post);
+    const textProvider = useSelector(({ ui }) => ui.textContent.projectPage);
+
+    const Paragraph = (params) => {
+        const sizeClass = `fs-${params?.size}`;
+        const weightClass = `fw-${params?.weight}`;
+
+        return (
+            <Grid item xs={12} className={internalClasses.sectionContent}>
+                <Typography variant="body1" className={clsx(sizeClass, weightClass)}>
+                    {params?.content}
+                </Typography>
+            </Grid>
+        );
+    };
+
+    const Image = ({ content }) => {
+        const sizeClass = content?.size === 'auto' ? internalClasses.imageSizeAuto : internalClasses.imageSizeNormal;
+
+        return (
+            <Grid item xs={12} className={internalClasses.sectionContent}>
+                <div className={clsx(internalClasses.image, sizeClass)}>
+                    {content?.src && <img src={content.src} alt={content?.alt} />}
+                </div>
+            </Grid>
+        );
+    };
+
+    const ImagesSlide = ({ content }) => {
+        return (
+            <Grid item xs={12} className={internalClasses.sectionContent}>
+                <div className={internalClasses.imagesSlideWrapper}>
+                    {Array.isArray(content) &&
+                        content.map((item) => (
+                            <div className={clsx(internalClasses.image, internalClasses.imageSizeNormal)}>
+                                {item?.src && <img src={item?.src} alt={item?.alt} />}
+                            </div>
+                        ))}
+                </div>
+            </Grid>
+        );
+    };
+
+    const Video = ({ content }) => {
+        return (
+            <Grid item xs={12} md={5} className={internalClasses.sectionContent}>
+                <div className={clsx(internalClasses.video)}>{content?.src && <video src={content.src} />}</div>
+            </Grid>
+        );
+    };
+
+    const renderProperElement = ({ type, ...rest }) => {
+        const itemOptions = {
+            image: Image,
+            video: Video,
+            paragraph: Paragraph,
+            'images-slide': ImagesSlide
+        };
+        const Component = itemOptions[type];
+
+        if (!Component) return null;
+
+        return <Component {...rest} />;
+    };
 
     return (
         <>
-            <Grid id="brief" spacing={4} container className={internalClasses.section} component="section" item xs={12}>
-                <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
-                    <Typography variant="body1" className="fs-biggest fw-800">
-                        01.
-                    </Typography>
-                </Grid>
-                <Grid item container xs={12} sm={9} lg={12}>
-                    <Grid item xs={12}>
-                        {Array.isArray(project?.brief) ? (
-                            project.brief.map((paragraph) => (
-                                <Typography
-                                    key={paragraph}
-                                    variant="body1"
-                                    className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                                >
-                                    {paragraph}
-                                </Typography>
-                            ))
-                        ) : (
-                            <Typography
-                                variant="body1"
-                                className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                            >
-                                {project?.brief}
-                            </Typography>
-                        )}
-                    </Grid>
-                    <Grid item xs={12}>
-                        <div className={clsx(internalClasses.image, internalClasses.firstSectionImage)}>
-                            {images[project?.id]?.image1 && (
-                                <img src={images[project.id].image1} alt={`${project?.id}1`} />
-                            )}
-                        </div>
-                    </Grid>
-                </Grid>
-            </Grid>
-            <Grid
-                id="research"
-                spacing={4}
-                container
-                className={internalClasses.section}
-                component="section"
-                item
-                xs={12}
-            >
-                <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
-                    <Typography variant="body1" className="fs-biggest fw-800">
-                        02.
-                    </Typography>
-                </Grid>
-                <Grid container item xs={12} sm={9} lg={12}>
-                    <Grid item xs={12}>
-                        {Array.isArray(project?.research) ? (
-                            project.research.map((paragraph) => (
-                                <Typography
-                                    key={paragraph}
-                                    variant="body1"
-                                    className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                                >
-                                    {paragraph}
-                                </Typography>
-                            ))
-                        ) : (
-                            <Typography
-                                variant="body1"
-                                className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                            >
-                                {project?.research}
-                            </Typography>
-                        )}
+            {project?.map((section) => (
+                <Grid
+                    key={section?.id}
+                    id={section?.id}
+                    spacing={4}
+                    container
+                    className={internalClasses.section}
+                    component="section"
+                    item
+                    xs={12}
+                >
+                    <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
+                        <Typography variant="body1" className="fs-biggest fw-800">
+                            {section?.sectionNumber}.
+                        </Typography>
                     </Grid>
 
-                    <Grid item xs={12}>
-                        <div className={clsx(internalClasses.image, internalClasses.secondSectionImage)}>
-                            {images[project?.id]?.image2 && (
-                                <img src={images[project.id].image2} alt={`${project?.id}2`} />
-                            )}
-                        </div>
+                    <Grid item container xs={12} sm={9} lg={12}>
+                        {section?.content?.map((item) => renderProperElement(item))}
                     </Grid>
                 </Grid>
-            </Grid>
-            <Grid
-                id="visualConcept"
-                spacing={4}
-                container
-                className={internalClasses.section}
-                component="section"
-                item
-                xs={12}
-            >
-                <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
-                    <Typography variant="body1" className="fs-biggest fw-800">
-                        03.
-                    </Typography>
-                </Grid>
-                <Grid item xs={12} sm={9} lg={12}>
-                    {Array.isArray(project?.visualConcept) ? (
-                        project.visualConcept.map((paragraph) => (
-                            <Typography
-                                key={paragraph}
-                                variant="body1"
-                                className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                            >
-                                {paragraph}
-                            </Typography>
-                        ))
-                    ) : (
-                        <Typography variant="body1" className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}>
-                            {project?.visualConcept}
-                        </Typography>
-                    )}
-                </Grid>
-                <Grid item xs={12} md={6}>
-                    <div className={clsx(internalClasses.image, internalClasses.thirdSectionImage)}>
-                        {images[project?.id]?.image5 && <img src={images[project.id].image5} alt={`${project?.id}5`} />}
-                    </div>
-                </Grid>
-                <Grid item xs={12} md={6}>
-                    <div className={clsx(internalClasses.image, internalClasses.thirdSectionImage)}>
-                        {images[project?.id]?.image6 && <img src={images[project.id].image6} alt={`${project?.id}6`} />}
-                    </div>
-                </Grid>
-                <Grid item xs={12} md={6}>
-                    <div className={clsx(internalClasses.image, internalClasses.thirdSectionImage)}>
-                        {images[project?.id]?.image7 && <img src={images[project.id].image7} alt={`${project?.id}7`} />}
-                    </div>
-                </Grid>
-                <Grid item xs={12} md={6}>
-                    <div className={clsx(internalClasses.image, internalClasses.thirdSectionImage)}>
-                        {images[project?.id]?.image8 && <img src={images[project.id].image8} alt={`${project?.id}8`} />}
-                    </div>
-                </Grid>
-            </Grid>
-            <Grid
-                id="prototype"
-                spacing={4}
-                container
-                className={internalClasses.section}
-                component="section"
-                item
-                xs={12}
-            >
-                <Grid item xs={12} sm={3} lg={false} className={internalClasses.sectionIndex}>
-                    <Typography variant="body1" className="fs-biggest fw-800">
-                        04.
-                    </Typography>
-                </Grid>
-                <Grid item xs={12} sm={9} lg={12}>
-                    {Array.isArray(project?.prototype) ? (
-                        project.prototype.map((paragraph) => (
-                            <Typography
-                                key={paragraph}
-                                variant="body1"
-                                className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}
-                            >
-                                {paragraph}
-                            </Typography>
-                        ))
-                    ) : (
-                        <Typography variant="body1" className={clsx(internalClasses.sectionContent, 'fs-200 fw-500')}>
-                            {project?.prototype}
-                        </Typography>
-                    )}
-                </Grid>
-                <Grid container item xs={12} justifyContent="center">
-                    <Grid item xs={12} md={5}>
-                        <div className={clsx(internalClasses.image, internalClasses.fourthSectionImage)}>
-                            {images[project?.id]?.image9 && (
-                                <img src={images[project.id].image9} alt={`${project?.id}9`} />
-                            )}
-                        </div>
-                    </Grid>
-                </Grid>
-            </Grid>
+            ))}
+
+            <Typography variant="body1" className={clsx(internalClasses.checkOutLink, 'fs-biggest fw-600')}>
+                {textProvider?.checkHereLink}{' '}
+                <a rel="noreferrer" target="_blank" href={project?.behanceLink} className={internalClasses?.hereLink}>
+                    {textProvider?.hereTitle}
+                </a>
+            </Typography>
         </>
     );
 };